+++ packages/saved-views-react/src/SavedViewsClient/ITwinSavedViewsClient.ts
@@ -3,15 +3,14 @@
 * See LICENSE.md in the project root for license terms and full copyright notice.
 *--------------------------------------------------------------------------------------------*/
 import {
-  ITwinSavedViewsClient as Client, type Group, type SavedViewMinimal, type SavedViewRepresentation,
-  type Tag,
+  ITwinSavedViewsClient as Client, type Group, type SavedViewRepresentation, type Tag,
 } from "@itwin/saved-views-client";
 
 import type { SavedView, SavedViewGroup, SavedViewTag } from "../SavedView.js";
 import type {
   CreateGroupParams, CreateSavedViewParams, CreateTagParams, DeleteGroupParams, DeleteSavedViewParams, DeleteTagParams,
-  GetSavedViewInfoParams, GetSingularSavedViewParams, GetThumbnailUrlParams, SavedViewInfo, SavedViewsClient,
-  UpdateGroupParams, UpdateSavedViewParams, UpdateTagParams, UploadThumbnailParams,
+  GetAllGroupsParams, GetAllSavedViewsParams, GetAllTagsParams, GetSavedViewParams, GetThumbnailUrlParams,
+  SavedViewsClient, UpdateGroupParams, UpdateSavedViewParams, UpdateTagParams, UploadThumbnailParams,
 } from "./SavedViewsClient.js";
 
 interface ITwinSavedViewsClientParams {
@@ -25,38 +24,44 @@ interface ITwinSavedViewsClientParams {
   baseUrl?: string | undefined;
 }
 
-/**  */
 export class ITwinSavedViewsClient implements SavedViewsClient {
-  private client: Client;
+  #client: Client;
 
   constructor(args: ITwinSavedViewsClientParams) {
-    this.client = new Client(args);
+    this.#client = new Client(args);
   }
 
-  public async getSavedViewInfo(args: GetSavedViewInfoParams): Promise<SavedViewInfo> {
-    const [{ savedViews }, { groups }, { tags }] = await Promise.all([
-      this.client.getAllSavedViewsMinimal(args),
-      this.client.getAllGroups(args),
-      this.client.getAllTags(args),
-    ]);
-
-    return {
-      savedViews: savedViews.map(savedViewResponseToSavedView),
-      groups: groups.map(groupResponseToSavedViewGroup),
-      tags: tags.map(tagResponseToSavedViewTag),
-    };
-  }
-
-  public async getSingularSavedView(args: GetSingularSavedViewParams): Promise<SavedViewRepresentation> {
-    const response = await this.client.getSavedViewRepresentation({
-      savedViewId: args.savedViewId,
+  async *getAllSavedViews(args: GetAllSavedViewsParams): AsyncIterableIterator<SavedView[]> {
+    const iterable = this.#client.getAllSavedViewsRepresentation({
+      iTwinId: args.iTwinId,
+      iModelId: args.iModelId,
       signal: args.signal,
     });
-    return response.savedView;
+    for await (const page of iterable) {
+      yield page.savedViews.map(savedViewResponseToSavedView);
+    }
   }
 
-  public async getThumbnailUrl(args: GetThumbnailUrlParams): Promise<string | undefined> {
-    const response = await this.client.getImage({
+  async getAllGroups(args: GetAllGroupsParams): Promise<SavedViewGroup[]> {
+    const response = await this.#client.getAllGroups({
+      iTwinId: args.iTwinId,
+      iModelId: args.iModelId,
+      signal: args.signal,
+    });
+    return response.groups.map(groupResponseToSavedViewGroup);
+  }
+
+  async getAllTags(args: GetAllTagsParams): Promise<SavedViewTag[]> {
+    const response = await this.#client.getAllTags({
+      iTwinId: args.iTwinId,
+      iModelId: args.iModelId,
+      signal: args.signal,
+    });
+    return response.tags.map(tagResponseToSavedViewTag);
+  }
+
+  async getThumbnailUrl(args: GetThumbnailUrlParams): Promise<string | undefined> {
+    const response = await this.#client.getImage({
       savedViewId: args.savedViewId,
       size: "thumbnail",
       signal: args.signal,
@@ -64,16 +69,23 @@ export class ITwinSavedViewsClient implements SavedViewsClient {
     return response.href;
   }
 
-  public async uploadThumbnail(args: UploadThumbnailParams): Promise<void> {
-    await this.client.updateImage({
+  async uploadThumbnail(args: UploadThumbnailParams): Promise<void> {
+    await this.#client.updateImage({
       savedViewId: args.savedViewId,
       image: args.image,
       signal: args.signal,
     });
   }
 
-  public async createSavedView(args: CreateSavedViewParams): Promise<SavedView> {
-    const { savedView } = await this.client.createSavedView({
+  async getSavedView(args: GetSavedViewParams): Promise<SavedView> {
+    const response = await this.#client.getSavedViewRepresentation({
+      savedViewId: args.savedViewId,
+      signal: args.signal,
+    });
+    return savedViewResponseToSavedView(response.savedView);
+  }
+  async createSavedView(args: CreateSavedViewParams): Promise<SavedView> {
+    const { savedView } = await this.#client.createSavedView({
       iTwinId: args.iTwinId,
       iModelId: args.iModelId,
       displayName: args.savedView.displayName,
@@ -87,8 +99,8 @@ export class ITwinSavedViewsClient implements SavedViewsClient {
     return savedViewResponseToSavedView(savedView);
   }
 
-  public async updateSavedView(args: UpdateSavedViewParams): Promise<SavedView> {
-    const { savedView } = await this.client.updateSavedView({
+  async updateSavedView(args: UpdateSavedViewParams): Promise<SavedView> {
+    const { savedView } = await this.#client.updateSavedView({
       savedViewId: args.savedView.id,
       displayName: args.savedView.displayName,
       tagIds: args.savedView.tagIds,
@@ -99,7 +111,7 @@ export class ITwinSavedViewsClient implements SavedViewsClient {
     });
 
     await Promise.all((args.savedView.extensions ?? []).map(
-      ({ extensionName, data }) => this.client.createExtension({
+      ({ extensionName, data }) => this.#client.createExtension({
         savedViewId: args.savedView.id,
         extensionName,
         data,
@@ -109,12 +121,12 @@ export class ITwinSavedViewsClient implements SavedViewsClient {
     return savedViewResponseToSavedView(savedView);
   }
 
-  public async deleteSavedView(args: DeleteSavedViewParams): Promise<void> {
-    await this.client.deleteSavedView({ savedViewId: args.savedViewId, signal: args.signal });
+  async deleteSavedView(args: DeleteSavedViewParams): Promise<void> {
+    await this.#client.deleteSavedView({ savedViewId: args.savedViewId, signal: args.signal });
   }
 
-  public async createGroup(args: CreateGroupParams): Promise<SavedViewGroup> {
-    const { group } = await this.client.createGroup({
+  async createGroup(args: CreateGroupParams): Promise<SavedViewGroup> {
+    const { group } = await this.#client.createGroup({
       iTwinId: args.iTwinId,
       iModelId: args.iModelId,
       displayName: args.group.displayName,
@@ -123,8 +135,8 @@ export class ITwinSavedViewsClient implements SavedViewsClient {
     return groupResponseToSavedViewGroup(group);
   }
 
-  public async updateGroup(args: UpdateGroupParams): Promise<SavedViewGroup> {
-    const { group } = await this.client.updateGroup({
+  async updateGroup(args: UpdateGroupParams): Promise<SavedViewGroup> {
+    const { group } = await this.#client.updateGroup({
       groupId: args.group.id,
       displayName: args.group.displayName,
       shared: args.group.shared,
@@ -133,16 +145,19 @@ export class ITwinSavedViewsClient implements SavedViewsClient {
     return groupResponseToSavedViewGroup(group);
   }
 
-  public async deleteGroup(args: DeleteGroupParams): Promise<void> {
-    const savedViews = await this.client.getAllSavedViewsMinimal({ groupId: args.groupId, signal: args.signal });
-    await Promise.all(
-      savedViews.savedViews.map(({ id }) => this.client.deleteSavedView({ savedViewId: id, signal: args.signal })),
-    );
-    await this.client.deleteGroup({ groupId: args.groupId, signal: args.signal });
+  async deleteGroup(args: DeleteGroupParams): Promise<void> {
+    const savedViewPages = this.#client.getAllSavedViewsMinimal({ groupId: args.groupId, signal: args.signal });
+    for await (const savedViews of savedViewPages) {
+      await Promise.all(
+        savedViews.savedViews.map(({ id }) => this.#client.deleteSavedView({ savedViewId: id, signal: args.signal })),
+      );
+    }
+
+    await this.#client.deleteGroup({ groupId: args.groupId, signal: args.signal });
   }
 
-  public async createTag(args: CreateTagParams): Promise<SavedViewTag> {
-    const { tag } = await this.client.createTag({
+  async createTag(args: CreateTagParams): Promise<SavedViewTag> {
+    const { tag } = await this.#client.createTag({
       iTwinId: args.iTwinId,
       iModelId: args.iModelId,
       displayName: args.displayName,
@@ -151,8 +166,8 @@ export class ITwinSavedViewsClient implements SavedViewsClient {
     return tagResponseToSavedViewTag(tag);
   }
 
-  public async updateTag(args: UpdateTagParams): Promise<SavedViewTag> {
-    const { tag } = await this.client.updateTag({
+  async updateTag(args: UpdateTagParams): Promise<SavedViewTag> {
+    const { tag } = await this.#client.updateTag({
       tagId: args.tag.id,
       displayName: args.tag.displayName,
       signal: args.signal,
@@ -160,22 +175,22 @@ export class ITwinSavedViewsClient implements SavedViewsClient {
     return tagResponseToSavedViewTag(tag);
   }
 
-  public async deleteTag(args: DeleteTagParams): Promise<void> {
-    await this.client.deleteTag({ tagId: args.tagId, signal: args.signal });
+  async deleteTag(args: DeleteTagParams): Promise<void> {
+    await this.#client.deleteTag({ tagId: args.tagId, signal: args.signal });
   }
 }
 
-function savedViewResponseToSavedView(response: Omit<SavedViewMinimal, "savedViewData">): SavedView {
+function savedViewResponseToSavedView(response: SavedViewRepresentation): SavedView {
   return {
     id: response.id,
     displayName: response.displayName,
-    viewData: undefined,
+    viewData: response.savedViewData,
     tagIds: response.tags?.map((tag) => tag.id),
     groupId: response._links.group?.href.split("/").at(-1),
     creatorId: response._links.creator?.href.split("/").at(-1),
     shared: response.shared,
     thumbnail: undefined,
-    extensions: undefined,
+    extensions: response.extensions,
     creationTime: response.creationTime,
     lastModified: response.lastModified,
   };