+++ packages/saved-views-react/src/SavedViewsClient/SavedViewsClient.ts
@@ -2,22 +2,16 @@
 * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
 * See LICENSE.md in the project root for license terms and full copyright notice.
 *--------------------------------------------------------------------------------------------*/
-import type { SavedViewRepresentation } from "@itwin/saved-views-client";
-
 import type { SavedView, SavedViewGroup, SavedViewTag, WriteableSavedViewProperties } from "../SavedView.js";
 import type { PartialExcept } from "../utils.js";
 
-export interface SavedViewInfo {
-  savedViews: SavedView[];
-  groups: SavedViewGroup[];
-  tags: SavedViewTag[];
-}
-
 export interface SavedViewsClient {
-  getSavedViewInfo: (args: GetSavedViewInfoParams) => Promise<SavedViewInfo>;
-  getSingularSavedView: (args: GetSingularSavedViewParams) => Promise<SavedViewRepresentation>;
+  getAllSavedViews: (args: GetAllSavedViewsParams) => AsyncIterableIterator<SavedView[]>;
+  getAllGroups: (args: GetAllGroupsParams) => Promise<SavedViewGroup[]>;
+  getAllTags: (args: GetAllTagsParams) => Promise<SavedViewTag[]>;
   getThumbnailUrl: (args: GetThumbnailUrlParams) => Promise<string | undefined>;
   uploadThumbnail: (args: UploadThumbnailParams) => Promise<void>;
+  getSavedView: (args: GetSavedViewParams) => Promise<SavedView>;
   createSavedView: (args: CreateSavedViewParams) => Promise<SavedView>;
   updateSavedView: (args: UpdateSavedViewParams) => Promise<SavedView>;
   deleteSavedView: (args: DeleteSavedViewParams) => Promise<void>;
@@ -29,13 +23,19 @@ export interface SavedViewsClient {
   deleteTag: (args: DeleteTagParams) => Promise<void>;
 }
 
-export interface GetSavedViewInfoParams extends CommonParams {
+export interface GetAllSavedViewsParams extends CommonParams {
   iTwinId: string;
   iModelId?: string | undefined;
 }
 
-export interface GetSingularSavedViewParams extends CommonParams {
-  savedViewId: string;
+export interface GetAllGroupsParams extends CommonParams {
+  iTwinId: string;
+  iModelId?: string | undefined;
+}
+
+export interface GetAllTagsParams extends CommonParams {
+  iTwinId: string;
+  iModelId?: string | undefined;
 }
 
 export interface GetThumbnailUrlParams extends CommonParams {
@@ -44,8 +44,17 @@ export interface GetThumbnailUrlParams extends CommonParams {
 
 export interface UploadThumbnailParams extends CommonParams {
   savedViewId: string;
-  /** Image data encoded as base64 data URL. */
+  /**
+   * Image data encoded as base64 data URL.
+   *
+   * @example
+   * "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAQAAAC1HAwCAAAAC0lEQVQYV2NgYAAAAAMAAWgmWQ0AAAAASUVORK5CYII="
+   */
   image: string;
+}
+
+export interface GetSavedViewParams extends CommonParams {
+  savedViewId: string;
 }
 
 export interface CreateSavedViewParams extends CommonParams {